First, render.
I have made a Card component that receives all elements from the data and renders a Card container.
At the receipts components after fetching the data from the API I am mapping throw the data array and rendering a card for each.
For a cleaner code, I have used spread syntax passing the props to the card component.
To avoid misleads and code complexity I am using TypeScript to fix and show required variables/props and their types for each section.
It helps reading and to understand the code.