Implementing a hover hook.
I am implementing a hover hook that helps me to add query selectors to any element required too.
Adding EventListeniers in one function and recalling it benefits the process greatly as we only set the event listeners in one place (once) and we remove them once.
So I do not need to rewrite a useEffect and its cleanup any in any component I need event listeners.