Animation applied to the status and the cards.
I decided to change the animation from width size to opacity as on width changing was changing the grid layout and it was so confusing.
However, the way the description cards appear now is more clear and not distracting.
Also, I highlighted the card hovered on by adding a box shadow to the card.