Screen Responsiveness finished.
I removed the header in mobile screens as it would not help the visibility of the system.
The grid layout for cards switched from row to column to fit the container better and conserve the visibility of the elements.
SnackBar position changed as on mobile screens it was blocking the thumbnail.