Use useCallback to marck the last element rendered on the list.
I am using useCallback to keep track of the reference to the last element rendered on the list.
The benefit useCallback provides is as useRef is not a part of the state that would not update when the last element changes, but useCallback in an observer hook-like useEffect and register the reference to the state.